requirements.txt
streamlit
langchain
PyPDF2
chromadb
sentence-transformers
transformers
python-dotenv
google-generativeai
